Update LSP keymaps and modify movement shortcuts
Implemented keymap configuration changes in the lsp.lua file to include additional keymaps for lsp. Disabled the "K" keymap and added a new keymap for "<leader>k". This is done to improve navigation and control in the lsp configuration. Also, modified navigation shortcuts in keymaps.lua, replacing "D" and "U" with "J" and "K" for more intuitive movement commands.